MQTT/requirements.txt
paho-mqtt==2.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
typing_extensions==4.12.2